geolocation/test_launcher.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Power API</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body/>

  <script type="application/javascript" >
  SimpleTest.waitForExplicitFinish();
  SpecialPowers.addPermission("settings", true, "http://example.org/");
  SpecialPowers.addPermission("geolocation", true, "http://example.org/");
  SpecialPowers.addPermission("browser", true, "http://example.org/");


  var iframe = document.createElement("iframe");
  iframe.mozApp = "http://example.org/manifest.webapp";
  iframe.src = "http://example.org/tests/dom/tests/mochitest/geolocation/allowCurrent.html";
  iframe.id = "ifrm";
  iframe.onload = loaded();
  iframe.mozbrowser = true;
  
  document.body.appendChild(iframe);

  function loaded() {

    window.addEventListener("message", function onMessage(event) {
       if (event.data == "finished") {
         window.removeEventListener("message", onMessage, false);
         SimpleTest.finish();
       } else {
         dump(JSON.stringify(event.data));
         ok(event.data.result, event.data.msg);
       }
           
     }, false);

  }
  </script>
</head>

</html>
